% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expense.R
\name{add_expense}
\alias{add_expense}
\title{Add an Expense to the Data Frame}
\usage{
add_expense(category, amount, expenses_data, selected_categories, session)
}
\arguments{
\item{amount}{The amount of the new expense, must be non-negative.}

\item{expenses_data}{A reactive value containing a data frame of expenses.}

\item{selected_categories}{A reactive value storing the list of already selected categories.}
}
\value{
Logical indicating whether the expense was successfully added (TRUE) or not (FALSE).
}
\description{
This function adds a new expense entry to the given data frame of expenses. It checks for the validity of the category and amount before adding.
}
\examples{
# In a Shiny server function:
expenses_data <- reactiveVal(data.frame(category = character(), amount = numeric()))
selected_categories <- reactiveVal(character())

observeEvent(input$add_expense, {
  result <- add_expense(input$category, input$expense, expenses_data, selected_categories)
  if (result) {
    # Update UI or perform other actions
  }
})
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
