% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_financials.R
\name{calculate_financials}
\alias{calculate_financials}
\title{Calculate Total Expenses and Savings}
\usage{
calculate_financials(expenses_data, income)
}
\arguments{
\item{expenses_data}{A data frame containing expenses.}

\item{income}{Numeric value of the user's income.}
}
\value{
A list containing the total expenses and savings.
}
\description{
This function calculates the total expenses and savings based on the provided data frame of expenses and the user's income.
}
\examples{
expenses <- data.frame(category = c("Food", "Rent"), amount = c(100, 500))
financials <- calculate_financials(expenses, 2000)
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
