% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\alias{generate_scatter_or_pie}
\title{Create a Scatter Plot of Expenses Data}
\usage{
generate_scatter_or_pie(
  expenses_data_summary,
  scatter_plot_type,
  colorblind_switch
)
}
\arguments{
\item{expenses_data}{A data frame containing expenses data.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Generates a scatter plot showing the distribution and percentage share of each expense category in the total expenses.
}
\examples{
expenses <- data.frame(category = c("Food", "Rent"), amount = c(100, 500))
scatter_plot <- create_scatter_plot(expenses)
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
