% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{createBarChart}
\alias{createBarChart}
\title{Create a Bar Chart with Plotly}
\usage{
createBarChart(
  data,
  title = "Income, Expenses, and Savings",
  xLabel = "Category",
  yLabel = "Amount"
)
}
\arguments{
\item{data}{A data frame containing the variables `category` and `amount`.}

\item{title}{Title of the plot, default is "Income, Expenses, and Savings".}

\item{xLabel}{Label for the x-axis, default is "Category".}

\item{yLabel}{Label for the y-axis, default is "Amount".}
}
\value{
An interactive Plotly object.
}
\description{
This function creates an interactive bar chart using ggplot2 and Plotly. It is designed to visualize
data in categories with their corresponding amounts. The chart uses a minimalistic color palette and
hides the legend for a cleaner appearance.
}
\examples{
test_data <- data.frame(category = c("Income", "Expenses", "Savings"),
                       amount = c(1000, 500, 500))
createBarChart(test_data)
}
