% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPieChart.R
\name{createPieChart}
\alias{createPieChart}
\title{Create a Pie Chart with Plotly}
\usage{
createPieChart(data, colorblind_switch = FALSE)
}
\arguments{
\item{data}{A data frame with at least two columns: category and value.}

\item{colorblind_switch}{A boolean flag to use colorblind-friendly palette. Default is FALSE.}
}
\value{
A ggplot object representing the pie chart.
}
\description{
This function takes a data frame with categories and their corresponding values and generates a pie chart. It optionally supports colorblind-friendly color palettes.
}
\examples{
data <- data.frame(
  category = c("Category 1", "Category 2", "Category 3"),
  value = c(10, 20, 30)
)
createPieChart(data, TRUE)
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
