% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scatter_or_pie.R
\name{generate_scatter_or_pie}
\alias{generate_scatter_or_pie}
\title{Generate Scatter Plot or Pie Chart}
\usage{
generate_scatter_or_pie(
  expenses_data_summary,
  scatter_plot_type,
  colorblind_switch
)
}
\arguments{
\item{expenses_data_summary}{A data frame summarizing expenses data. It should contain at least two columns: 'category' for the expense categories and 'percentage' for the percentage share of each category.}

\item{scatter_plot_type}{A character string specifying the type of plot to generate. Expected values are "Scatter Plot" or "Pie Chart".}

\item{colorblind_switch}{A logical value indicating whether to use a colorblind-friendly palette.}
}
\value{
A Plotly object representing the specified type of plot (scatter plot or pie chart).
}
\description{
This function creates either a scatter plot or a pie chart based on user input. It is designed to be used within a Shiny application.
}
\examples{
# Example usage within a Shiny server function:
expenses_summary <- data.frame(
  category = c("Food", "Transport", "Utilities"),
  amount = c(200, 150, 100),
  percentage = c(40, 30, 30)
)
output$myPlot <- renderPlotly({
  generate_scatter_or_pie(expenses_summary, input$plotType, input$colorblindSwitch)
})
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
