% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMyApp.R
\name{launchMyApp}
\alias{launchMyApp}
\title{Launch My Shiny App}
\usage{
launchMyApp()
}
\description{
This function starts the Shiny application contained in the package.
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
