% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_function.R
\name{runDemo}
\alias{runDemo}
\title{Demonstration of shiny app}
\usage{
runDemo()
}
\value{
Open the app of Budget Overview Shiny Dashboard
}
\description{
This function is used to demonstrate how the shiny app works.
}
\details{
The `runDemo` function uses `shiny::addResourcePath` to set the resource path to the package's internal `styles` directory, ensuring that Shiny can access the CSS file upon deployment. It then includes a link to the CSS file in the app's HTML header using `tags$link`. This setup allows for custom styling of the Shiny dashboard elements.
}
\examples{
\dontrun{runDemo()}
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
