% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expense.R
\name{add_expense}
\alias{add_expense}
\title{Add an Expense to the Data Frame}
\usage{
add_expense(expenses_data, category, amount)
}
\arguments{
\item{expenses_data}{A data frame containing existing expenses.}

\item{category}{The category of the new expense.}

\item{amount}{The amount of the new expense.}
}
\value{
A data frame with the new expense added.
}
\description{
This function adds a new expense entry to the given data frame of expenses. It checks for the validity of the category and amount before adding.
}
\examples{
expenses <- data.frame(category = character(), amount = numeric())
expenses <- add_expense(expenses, "Food", 100)
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
