% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expense.R
\name{add_expense}
\alias{add_expense}
\title{Add an Expense to the Data Frame}
\usage{
add_expense(category, amount, expenses_data, selected_categories, session)
}
\arguments{
\item{category}{A string specifying the category of the expense.}

\item{amount}{A numeric value representing the amount of the expense. Must be non-negative.}

\item{expenses_data}{A reactive value containing a data frame of expenses.}

\item{selected_categories}{A reactive value storing the list of already selected categories.}

\item{session}{The Shiny session object passed to showModal for displaying modals.}
}
\value{
Logical TRUE if the expense is successfully added, FALSE otherwise.
}
\description{
This function adds a new expense entry to the given data frame of expenses. It checks for the validity of the category and amount before adding.
}
\examples{
\dontrun{
  # Simulate the environment of a Shiny app
  library(shiny)
  expenses_data <- reactiveVal(data.frame(category = character(), amount = numeric()))
  selected_categories <- reactiveVal(character())
  session <- shiny::getDefaultReactiveDomain()

  # Examples of add_expense function
  add_expense("Food", 50, expenses_data, selected_categories, session)
  add_expense("Transport", -10, expenses_data, selected_categories, session) # Invalid amount
  add_expense("Food", 30, expenses_data, selected_categories, session) # Category already exists
}
}
\author{
Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.
}
