% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot_f.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Create a Scatter Plot of Expenses Data}
\usage{
create_scatter_plot(expenses_data)
}
\arguments{
\item{expenses_data}{A data frame containing expenses data.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Generates a scatter plot showing the distribution and percentage share of
each expense category in the total expenses.
}
\examples{
expenses <- data.frame(category = c("Food", "Rent"), amount = c(100, 500))
scatter_plot <- create_scatter_plot(expenses)
}
